WIP: update custom color examples
