+++ stories/examples/custom-bpmn-theme/custom-colors/CustomColorForBPMNKind.js
@@ -5,10 +5,9 @@ import '../../static/css/icons.min.css';
 import '../../static/css/main.css';
 
 import { getCustomColorsBpmnDiagram } from '../../static/js/diagram/bpmn-diagrams';
-import { StyleIdentifiers, Directions, FontStyle } from "../../static/js/style-identifiers";
+import {Directions, FontStyle, StyleIdentifiers} from "../../static/js/style-identifiers";
 
 import { BpmnVisualization, ShapeUtil, ShapeBpmnElementKind } from 'bpmn-visualization';
-
 
 class BpmnVisualizationCustomizedColors extends BpmnVisualization {
   constructor(containerId) {
@@ -46,25 +45,6 @@ class BpmnVisualizationCustomizedColors extends BpmnVisualization {
     // for more details about font, see the custom-fonts example
     poolStyle[StyleIdentifiers.STYLE_FONTCOLOR] = 'White';
     poolStyle[StyleIdentifiers.STYLE_FONTSTYLE] = FontStyle.FONT_BOLD + FontStyle.FONT_ITALIC;
-
-
-    const startEventStyle = styleSheet.styles[ShapeBpmnElementKind.EVENT_START];
-    startEventStyle[StyleIdentifiers.STYLE_FILLCOLOR] = '#d6eea5';
-    startEventStyle[StyleIdentifiers.STYLE_STROKECOLOR] = '#8dc125';
-
-    [ShapeBpmnElementKind.EVENT_INTERMEDIATE_CATCH, ShapeBpmnElementKind.EVENT_INTERMEDIATE_THROW].forEach(kind => {
-      const intermediateEventStyle = styleSheet.styles[kind];
-      intermediateEventStyle[StyleIdentifiers.STYLE_STROKECOLOR] = '#7307df';
-      intermediateEventStyle[StyleIdentifiers.STYLE_FILLCOLOR] = 'white'; // ensure reset if the style is redefined before
-    })
-
-    const boundaryEventStyle = styleSheet.styles[ShapeBpmnElementKind.EVENT_BOUNDARY];
-    boundaryEventStyle[StyleIdentifiers.STYLE_FILLCOLOR] = 'LightGoldenrodYellow';
-    boundaryEventStyle[StyleIdentifiers.STYLE_STROKECOLOR] = 'DarkOrange';
-
-    const endEventStyle = styleSheet.styles[ShapeBpmnElementKind.EVENT_END];
-    endEventStyle[StyleIdentifiers.STYLE_FILLCOLOR] = 'Pink';
-    endEventStyle[StyleIdentifiers.STYLE_STROKECOLOR] = 'FireBrick';
   }
 }
 
@@ -75,7 +55,7 @@ function createElementsInDOM() {
   return container;
 }
 
-export const createCustomColorForBPMNKindAndSpecificElement = () => {
+export const createCustomColorForBPMNKind = () => {
   const container = createElementsInDOM();
 
   const bpmnVisualization = new BpmnVisualizationCustomizedColors({ container });
